Changed thread percentage to threads. Improved readme.
Changed the usage of the -p flag to -t and instead of giving a percentage
of threads for ffmpeg to use, instead give a number of threads. This value
isn't checked and so could be dangerous, but this script is just for personal
usage so it should be fine.

Also improved and updated the README to match new usage, and added installation
instructions using the improved executable script method instead of
recommending adding an alias.