Wasn't renaming the files after converting.
The files were being re-encoded to the new format, but only as a
container, the extension was kept the same because I forgot to
alter the file_new variable. Isn't an issue for programs like VLC
or on Linux where the programs will check the container, but windows
is bad.